+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * [Kernel] Compiled files now follow "__MAIN__-ModuleName" convention to solve issues with Erlang embedded mode;
 
 * deprecations
+  * [Kernel] loop and recur are deprecated in favor of recursion with named functions;
   * [Access] The semantics of the access protocol were reduced from a broad query API to simple data structure key-based access;
   * [Module] `Module.add_compile_callback(module, target, callback)` was deprecated in favor of the simpler `Module.add_attribute(module, :before_compile, { target, callback })` API;
   * [File] `File.read_info` was deprecated in favor of `File.stat`;